requirements.txt
openai>=1.30.0
duckdb>=1.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
prometheus_client>=0.20.0

# Dev / test
pytest>=8.0.0
httpx>=0.27.0
